+++ R/ddfilter.R
@@ -32,7 +32,7 @@
 #' "pTime" and "sTime" are hours from a previous and to a subsequent fix respectively. 
 #' "pDist" and "sDist" are straight distances in kilometres from a previous and to a subsequent fix respectively. 
 #' "pSpeed" and "sSpeed" are linear speed from a previous and to a subsequent fix respectively. 
-#' "inAng" is the angle between the bearings of lines joining successive location points.    
+#' "inAng" is the degree between the bearings of lines joining successive location points.
 #' @author Takahiro Shimada
 #' @references Shimada T, Jones R, Limpus C, Hamann M (2012) 
 #' Improving data retention and home range estimates by data-driven screening.