+++ R/vmax.R
@@ -43,36 +43,38 @@ vmax<-function(sdata, qi=5, prob=0.99){
   IDs<-levels(factor(sdata$id))
   
   
-  ## Hours from a previous and to a subsequent location (pTime & sTime)
-  stepTime<-function(j){
-      timeDiff<-diff(sdata[sdata$id %in% j, "DateTime"])
-      units(timeDiff)<-"hours"
-      c(as.numeric(timeDiff), NA)
-  } 
-  
-  sTime<-unlist(lapply(IDs, stepTime))  
-  sdata$pTime<-c(NA, sTime[-length(sTime)])
-  sdata$sTime<-sTime
-  
-  
-  ## Distance from a previous and to a subsequent location (pDist & sDist)
-  for(j in IDs){
-    # CUrrent location as "SpatialPoints"
-    turtle<-sdata[sdata$id %in% j,]  
-    LatLong<-data.frame(Y=turtle$lat, X=turtle$lon)
-    sp::coordinates(LatLong)<-~X+Y
-    sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
-    
-    #pDist
-    sdata[sdata$id %in% j,"pDist"]<-c(NA, raster::pointDistance(LatLong[-length(LatLong)], LatLong[-1], lonlat=T)/1000)
-    
-    #sDist
-    sdata[sdata$id %in% j,"sDist"]<-c(raster::pointDistance(LatLong[-1], LatLong[-length(LatLong)], lonlat=T)/1000,NA)
-  }     
-  
-  
-  # Speed from a previous and to a subsequent location in km/h
-  sdata$pSpeed<-sdata$pDist/sdata$pTime
+  # ## Hours from a previous and to a subsequent location (pTime & sTime)
+  # stepTime<-function(j){
+  #     timeDiff<-diff(sdata[sdata$id %in% j, "DateTime"])
+  #     units(timeDiff)<-"hours"
+  #     c(as.numeric(timeDiff), NA)
+  # } 
+  # 
+  # sTime<-unlist(lapply(IDs, stepTime))  
+  # sdata$pTime<-c(NA, sTime[-length(sTime)])
+  # sdata$sTime<-sTime
+  # 
+  # 
+  # ## Distance from a previous and to a subsequent location (pDist & sDist)
+  # for(j in IDs){
+  #   # CUrrent location as "SpatialPoints"
+  #   turtle<-sdata[sdata$id %in% j,]  
+  #   LatLong<-data.frame(Y=turtle$lat, X=turtle$lon)
+  #   sp::coordinates(LatLong)<-~X+Y
+  #   sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
+  #   
+  #   #pDist
+  #   sdata[sdata$id %in% j,"pDist"]<-c(NA, raster::pointDistance(LatLong[-length(LatLong)], LatLong[-1], lonlat=T)/1000)
+  #   
+  #   #sDist
+  #   sdata[sdata$id %in% j,"sDist"]<-c(raster::pointDistance(LatLong[-1], LatLong[-length(LatLong)], lonlat=T)/1000,NA)
+  # }     
+  # 
+  # 
+  # # Speed from a previous and to a subsequent location in km/h
+  # sdata$pSpeed<-sdata$pDist/sdata$pTime
+  # 
+  sdata <- track_param(sdata, param = c('time', 'distance', 'speed'))
   speed<-sdata$pSpeed[!(is.na(sdata$pSpeed))]