+++ R/vmaxlp.R
@@ -46,45 +46,45 @@ vmaxlp<-function(sdata, qi=4, prob=0.99){
   IDs<-levels(factor(sdata$id))
   
   
-  ## Hours from a previous and to a subsequent location (pTime & sTime)
-  stepTime<-function(j){
-    timeDiff<-diff(sdata[sdata$id %in% j, "DateTime"])
-    units(timeDiff)<-"hours"
-    c(as.numeric(timeDiff), NA)
-  } 
+  # ## Hours from a previous and to a subsequent location (pTime & sTime)
+  # stepTime<-function(j){
+  #   timeDiff<-diff(sdata[sdata$id %in% j, "DateTime"])
+  #   units(timeDiff)<-"hours"
+  #   c(as.numeric(timeDiff), NA)
+  # } 
+  # 
+  # sTime<-unlist(lapply(IDs, stepTime))  
+  # sdata$pTime<-c(NA, sTime[-length(sTime)])
+  # sdata$sTime<-sTime
+  # 
+  # 
+  # ## Distance from a previous and to a subsequent location (pDist & sDist)
+  # calcDist<-function(j){
+  #     turtle<-sdata[sdata$id %in% j,]  
+  #     LatLong<-data.frame(Y=turtle$lat, X=turtle$lon)
+  #     sp::coordinates(LatLong)<-~X+Y
+  #     sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
+  #     
+  #     #pDist
+  #     c(NA, raster::pointDistance(LatLong[-length(LatLong)], LatLong[-1], lonlat=T)/1000)
+  # }
+  # 
+  # sdata$pDist<-unlist(lapply(IDs, calcDist))
+  # sdata$sDist<-c(sdata$pDist[-1], NA)
+  # 
+  # 
+  # # Speed from a previous and to a subsequent location in km/h
+  # sdata$pSpeed<-sdata$pDist/sdata$pTime
+  # sdata$sSpeed<-sdata$sDist/sdata$sTime
   
-  sTime<-unlist(lapply(IDs, stepTime))  
-  sdata$pTime<-c(NA, sTime[-length(sTime)])
-  sdata$sTime<-sTime
-  
-  
-  ## Distance from a previous and to a subsequent location (pDist & sDist)
-  calcDist<-function(j){
-      turtle<-sdata[sdata$id %in% j,]  
-      LatLong<-data.frame(Y=turtle$lat, X=turtle$lon)
-      sp::coordinates(LatLong)<-~X+Y
-      sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
-      
-      #pDist
-      c(NA, raster::pointDistance(LatLong[-length(LatLong)], LatLong[-1], lonlat=T)/1000)
-  }
-  
-  sdata$pDist<-unlist(lapply(IDs, calcDist))
-  sdata$sDist<-c(sdata$pDist[-1], NA)
-  
-  
-  # Speed from a previous and to a subsequent location in km/h
-  sdata$pSpeed<-sdata$pDist/sdata$pTime
-  sdata$sSpeed<-sdata$sDist/sdata$sTime
-  
-  
-  ## Calculate inner angle in degree
-  LatLong<-data.frame(Y=sdata$lat, X=sdata$lon, tms=sdata$DateTime, id=sdata$id)
-  sp::coordinates(LatLong)<-~X+Y
-  sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
-  tr<-trip::trip(LatLong, c("tms", "id"))
-  sdata$inAng<-trip::trackAngle(tr)
-  sdata<-with(sdata, sdata[complete.cases(inAng),])
+  # ## Calculate inner angle in degree
+  # LatLong<-data.frame(Y=sdata$lat, X=sdata$lon, tms=sdata$DateTime, id=sdata$id)
+  # sp::coordinates(LatLong)<-~X+Y
+  # sp::proj4string(LatLong)<-sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
+  # tr<-trip::trip(LatLong, c("tms", "id"))
+  # sdata$inAng<-trip::trackAngle(tr)
+  # sdata<-with(sdata, sdata[complete.cases(inAng),])
+  sdata <- track_param(sdata, param = c('time', 'distance', 'speed', 'angle'))
   
   
   #### Exclude datasets less than 6 locations
@@ -225,20 +225,21 @@ vmaxlp<-function(sdata, qi=4, prob=0.99){
   row.names(sdata)<-1:nrow(sdata)
   
   
-  #### Calculate speed
-  sTime<-unlist(lapply(IDs, stepTime))  
-  sdata$pTime<-c(NA, sTime[-length(sTime)])
-  sdata$sTime<-sTime
-  
-  
-  ## Distance from a previous and to a subsequent location (pDist & sDist)
-  sdata$pDist<-unlist(lapply(IDs, calcDist))
-  sdata$sDist<-c(sdata$pDist[-1], NA)
-  
-  
-  ## Speed from a previous and to a subsequent location in km/h
-  sdata$pSpeed<-sdata$pDist/sdata$pTime
-  sdata$sSpeed<-sdata$sDist/sdata$sTime
+  # #### Calculate speed
+  # sTime<-unlist(lapply(IDs, stepTime))  
+  # sdata$pTime<-c(NA, sTime[-length(sTime)])
+  # sdata$sTime<-sTime
+  # 
+  # 
+  # ## Distance from a previous and to a subsequent location (pDist & sDist)
+  # sdata$pDist<-unlist(lapply(IDs, calcDist))
+  # sdata$sDist<-c(sdata$pDist[-1], NA)
+  # 
+  # 
+  # ## Speed from a previous and to a subsequent location in km/h
+  # sdata$pSpeed<-sdata$pDist/sdata$pTime
+  # sdata$sSpeed<-sdata$sDist/sdata$sTime
+  sdata <- track_param(sdata, param = c('time', 'distance', 'speed'))
   
   
   #### Retain locations with more than two consecutive points